requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.6
python-json-logger==2.0.7 